feat: ModeMan 크롤링 배치 잡 리팩토링 및 URL 처리 개선

+++ src/main/java/com/jj/redline/batch/output/MetaReport.java
@@ -22,10 +22,10 @@ public class MetaReport {
     private OffsetDateTime finishedAt;  // UTC 권장
     private Long durationMs;
 
-    private Integer total;
-    private Integer ok;
-    private Integer partial;
-    private Integer fail;
+    private Long total;
+    private Long ok;
+    private Long partial;
+    private Long fail;
 
     // 실패/부분실패 샘플(너무 길어지면 앞에서 N개만 담는 용도)
     private List<ErrorSample> errorSamples;
